+++ sickrage/core/webserver/views.py
@@ -114,14 +114,12 @@ class BaseHandler(RequestHandler):
                 url = url[len(sickrage.app.config.web_root) + 1:]
 
             if url[:3] != 'api':
-                return self.finish(self.render(
+                self.write(self.render(
                     '/errors/404.mako',
                     title=_('HTTP Error 404'),
-                    header=_('HTTP Error 404'))
-                )
+                    header=_('HTTP Error 404')))
             else:
                 self.write('Wrong API key used')
-
         elif self.settings.get("debug") and "exc_info" in kwargs:
             exc_info = kwargs["exc_info"]
             trace_info = ''.join(["%s<br>" % line for line in traceback.format_exception(*exc_info)])
@@ -131,21 +129,21 @@ class BaseHandler(RequestHandler):
 
             self.set_header('Content-Type', 'text/html')
             self.write("""<html>
-                                 <title>{error}</title>
-                                 <body>
-                                    <button onclick="window.location='{webroot}/logs/';">View Log(Errors)</button>
-                                    <button onclick="window.location='{webroot}/home/restart?force=1';">Restart SiCKRAGE</button>
-                                    <h2>Error</h2>
-                                    <p>{error}</p>
-                                    <h2>Traceback</h2>
-                                    <p>{traceback}</p>
-                                    <h2>Request Info</h2>
-                                    <p>{request}</p>
-                                 </body>
-                               </html>""".format(error=error,
-                                                 traceback=trace_info,
-                                                 request=request_info,
-                                                 webroot=sickrage.app.config.web_root))
+                             <title>{error}</title>
+                             <body>
+                                <button onclick="window.location='{webroot}/logs/';">View Log(Errors)</button>
+                                <button onclick="window.location='{webroot}/home/restart?force=1';">Restart SiCKRAGE</button>
+                                <h2>Error</h2>
+                                <p>{error}</p>
+                                <h2>Traceback</h2>
+                                <p>{traceback}</p>
+                                <h2>Request Info</h2>
+                                <p>{request}</p>
+                             </body>
+                           </html>""".format(error=error,
+                                             traceback=trace_info,
+                                             request=request_info,
+                                             webroot=sickrage.app.config.web_root))
 
     def get_current_user(self):
         user = self.get_secure_cookie('sr_user')
@@ -198,7 +196,7 @@ class BaseHandler(RequestHandler):
         return self.render_string(template_name, **kwargs)
 
     @run_on_executor
-    def route(self, function, **kwargs):
+    def worker(self, function, **kwargs):
         threading.currentThread().setName("TORNADO")
         kwargs = recursive_unicode(kwargs)
         for arg, value in kwargs.items():
@@ -229,17 +227,24 @@ class WebHandler(BaseHandler):
 
     @coroutine
     @authenticated
-    def prepare(self, *args, **kwargs):
+    def get(self, *args, **kwargs):
+        result = yield self.route()
+        if result: self.write(result)
+
+    @coroutine
+    @authenticated
+    def post(self, *args, **kwargs):
+        result = yield self.route()
+        if result: self.write(result)
+
+    def route(self):
         # route -> method obj
         method = getattr(
             self, self.request.path.strip('/').split('/')[::-1][0].replace('.', '_'),
             getattr(self, 'index', None)
         )
 
-        if method:
-            result = yield self.route(method, **self.request.arguments)
-            if self.request.method == 'GET' or result:
-                self.finish(result)
+        if method: return self.worker(method, **self.request.arguments)
 
     def _genericMessage(self, subject, message):
         return self.render(
@@ -290,7 +295,6 @@ class LogoutHandler(BaseHandler):
 
         self.clear_all_cookies()
         return self.redirect('/login/')
-
 
 class CalendarHandler(BaseHandler):
     def prepare(self, *args, **kwargs):
@@ -523,32 +527,6 @@ class WebRoot(WebHandler):
     def quicksearch_json(self, term):
         return json_encode(
             sickrage.app.quicksearch_cache.get_shows(term) + sickrage.app.quicksearch_cache.get_episodes(term))
-
-
-@Route('/ui(/?.*)')
-class UI(WebHandler):
-    def __init__(self, *args, **kwargs):
-        super(UI, self).__init__(*args, **kwargs)
-        self.set_header('Content-Type', 'application/json')
-
-    @staticmethod
-    def add_message():
-        sickrage.app.alerts.message('Test 1', 'This is test number 1')
-        sickrage.app.alerts.error('Test 2', 'This is test number 2')
-        return "ok"
-
-    def get_messages(self):
-        messages = {}
-        cur_notification_num = 0
-        for cur_notification in sickrage.app.alerts.get_notifications(self.request.remote_ip):
-            cur_notification_num += 1
-            messages['notification-{}'.format(cur_notification_num)] = {
-                'title': cur_notification.title,
-                'message': cur_notification.message or "",
-                'type': cur_notification.type
-            }
-
-        return json_encode(messages)
 
 
 @Route('/browser(/?.*)')